+++ freight/__init__.py
@@ -1,3 +1,3 @@
 default_app_config = 'freight.apps.FreightConfig'
 
-__version__ = '0.9.3'
+__version__ = '0.10.0'