+++ freight/tests.py
@@ -1,18 +1,22 @@
+import datetime
 import logging
 import inspect
 import json
 import math
 import os
+from random import randrange
 import sys
 from unittest.mock import Mock, patch
 
 from django.contrib.auth.models import User, Permission 
-from django.urls import reverse
 from django.test import TestCase
 from django.test.client import Client
+from django.urls import reverse
+from django.utils.timezone import now
 
 from allianceauth.eveonline.models import EveCharacter, EveCorporationInfo
 from allianceauth.authentication.models import CharacterOwnership
+from allianceauth.services.modules.discord.models import DiscordUser
 from esi.models import Token, Scope
 from esi.errors import TokenExpiredError, TokenInvalidError
 
@@ -24,7 +28,7 @@ from .templatetags.freight_filters import power10, formatnumber
 
 # reconfigure logger so we get logging from tasks to console during test
 c_handler = logging.StreamHandler(sys.stdout)
-logger = logging.getLogger('freight.tasks')
+logger = logging.getLogger('freight.managers')
 logger.level = logging.DEBUG
 logger.addHandler(c_handler)
 
@@ -200,9 +204,12 @@ class TestRunContractsSync(TestCase):
             'r', 
             encoding='utf-8'
         ) as f:
-            characters = json.load(f)
+            cls.characters_data = json.load(f)
 
-        for character in characters:
+    
+    def setUp(self):
+
+        for character in self.characters_data:
             EveCharacter.objects.create(**character)
             EveCorporationInfo.objects.get_or_create(
                 corporation_id=character['corporation_id'],
@@ -212,33 +219,30 @@ class TestRunContractsSync(TestCase):
                     'member_count': 42
                 }
             )
-
-        # setup test data
-        cls.client = Client()
-
-        # 1 user
-        cls.character = EveCharacter.objects.get(character_id=90000001)
         
-        cls.alliance = EveOrganization.objects.create(
-            id = cls.character.alliance_id,
+        # 1 user
+        self.character = EveCharacter.objects.get(character_id=90000001)
+        
+        self.alliance = EveOrganization.objects.create(
+            id = self.character.alliance_id,
             category = EveOrganization.CATEGORY_ALLIANCE,
-            name = cls.character.alliance_name
+            name = self.character.alliance_name
         )
-        cls.corporation = EveOrganization.objects.create(
-            id = cls.character.corporation_id,
+        self.corporation = EveOrganization.objects.create(
+            id = self.character.corporation_id,
             category = EveOrganization.CATEGORY_CORPORATION,
-            name = cls.character.corporation_name
+            name = self.character.corporation_name
         )
-        cls.user = User.objects.create_user(
-            cls.character.character_name,
+        self.user = User.objects.create_user(
+            self.character.character_name,
             'abc@example.com',
             'password'
         )
 
-        cls.main_ownership = CharacterOwnership.objects.create(
-            character=cls.character,
+        self.main_ownership = CharacterOwnership.objects.create(
+            character=self.character,
             owner_hash='x1',
-            user=cls.user
+            user=self.user
         )        
 
         # Locations
@@ -257,7 +261,6 @@ class TestRunContractsSync(TestCase):
             category_id=65
         )      
         
-
 
     # identify wrong operation mode
     @patch(
@@ -738,6 +741,7 @@ class TestRunContractsSync(TestCase):
         print('hi')
     """
 
+
 class TestFilters(TestCase):
 
     def test_power10(self):
@@ -807,3 +811,267 @@ class TestFilters(TestCase):
             formatnumber(None),
             None
         )
+
+
+class TestNotifications(TestCase):
+    
+    # note: setup is making calls to ESI to get full info for entities
+    # all ESI calls in the tested module are mocked though
+
+
+    @classmethod
+    def setUpClass(cls):
+        super(TestNotifications, cls).setUpClass()
+
+        # load test data
+        currentdir = os.path.dirname(os.path.abspath(inspect.getfile(
+            inspect.currentframe()
+        )))
+
+        # ESI contracts        
+        with open(
+            currentdir + '/testdata/contracts.json', 
+            'r', 
+            encoding='utf-8'
+        ) as f:
+            cls.contracts = json.load(f)
+                
+        # update dates to something current, so won't be treated as stale
+        for contract in cls.contracts:
+            date_issued = now() - datetime.timedelta(
+                days=randrange(5), 
+                hours=randrange(10)
+            )
+            date_accepted = date_issued + datetime.timedelta(
+                hours=randrange(5),
+                minutes=randrange(30)
+            )
+            date_completed = date_accepted + datetime.timedelta(
+                hours=randrange(12),
+                minutes=randrange(30)
+            )
+            date_expired = now() + datetime.timedelta(
+                days=randrange(14), 
+                hours=randrange(10)
+            )
+            if 'date_issued' in contract:
+                contract['date_issued'] = date_issued.isoformat()
+
+            if 'date_accepted' in contract:
+                contract['date_accepted'] = date_accepted.isoformat()
+
+            if 'date_completed' in contract:
+                contract['date_completed'] = date_completed.isoformat()
+
+            if 'date_expired' in contract:
+                contract['date_expired'] = date_expired.isoformat()
+            
+
+        # Eve characters
+        with open(
+            currentdir + '/testdata/characters.json', 
+            'r', 
+            encoding='utf-8'
+        ) as f:
+            cls.characters_data = json.load(f)
+
+    
+    def setUp(self):
+
+        for character in self.characters_data:
+            EveCharacter.objects.create(**character)
+            EveCorporationInfo.objects.get_or_create(
+                corporation_id=character['corporation_id'],
+                defaults={
+                    'corporation_name': character['corporation_name'],
+                    'corporation_ticker': character['corporation_ticker'],
+                    'member_count': 42
+                }
+            )
+        
+        # 1 user
+        self.character = EveCharacter.objects.get(character_id=90000001)
+        
+        self.organization = EveOrganization.objects.create(
+            id = self.character.alliance_id,
+            category = EveOrganization.CATEGORY_ALLIANCE,
+            name = self.character.alliance_name
+        )
+        
+        self.user = User.objects.create_user(
+            self.character.character_name,
+            'abc@example.com',
+            'password'
+        )
+
+        self.main_ownership = CharacterOwnership.objects.create(
+            character=self.character,
+            owner_hash='x1',
+            user=self.user
+        )        
+
+        # Locations
+        jita = Location.objects.create(
+            id=60003760,
+            name='Jita IV - Moon 4 - Caldari Navy Assembly Plant',
+            solar_system_id=30000142,
+            type_id=52678,
+            category_id=3
+        )
+        amamake = Location.objects.create(
+            id=1022167642188,
+            name='Amamake - 3 Time Nearly AT Winners',
+            solar_system_id=30002537,
+            type_id=35834,
+            category_id=65
+        )      
+
+        # create contracts
+        pricing = Pricing.objects.create(
+            start_location=jita,
+            end_location=amamake,
+            price_base=500000000
+        )
+        
+        handler = ContractHandler.objects.create(
+            organization=self.organization,
+            character=self.main_ownership            
+        )
+        
+        for contract in self.contracts:
+            Contract.objects.update_or_create_from_dict(
+                handler=handler,
+                contract=contract,
+                esi_client=Mock()
+            )
+
+        Contract.objects.update_pricing() 
+
+        # create users and Discord accounts from contract issuers
+        for contract in Contract.objects.all():
+            issuer_user = User.objects\
+                .filter(
+                    character_ownerships__character__exact=contract.issuer
+                )\
+                .first()
+            if not issuer_user:
+                user = User.objects.create_user(
+                    contract.issuer.character_name,
+                    'abc@example.com',
+                    'password'
+                )
+                CharacterOwnership.objects.create(
+                    character=contract.issuer,
+                    owner_hash=contract.issuer.character_name + 'x',
+                    user=user
+                )   
+            DiscordUser.objects.update_or_create(
+                user=user,
+                defaults={
+                    "uid": contract.issuer.character_id
+                }
+            )
+
+
+    @patch('freight.managers.FREIGHT_DISCORD_WEBHOOK_URL', 'url')
+    @patch('freight.managers.FREIGHT_DISCORD_CUSTOMERS_WEBHOOK_URL', None)
+    @patch('freight.models.Webhook.execute', autospec=True)
+    def test_send_pilot_notifications_normal(
+        self, 
+        mock_webhook_execute
+    ):        
+        self.assertTrue(tasks.send_contract_notifications(rate_limted=False))
+        self.assertEqual(mock_webhook_execute.call_count, 7)
+
+
+    @patch('freight.managers.FREIGHT_DISCORD_WEBHOOK_URL', None)
+    @patch('freight.managers.FREIGHT_DISCORD_CUSTOMERS_WEBHOOK_URL', 'url')    
+    @patch('freight.models.Webhook.execute', autospec=True)
+    def test_send_customer_notifications_normal(
+        self, 
+        mock_webhook_execute
+    ):        
+        self.assertTrue(tasks.send_contract_notifications(rate_limted=False))
+        self.assertEqual(mock_webhook_execute.call_count, 10)
+
+    
+    @patch('freight.managers.FREIGHT_DISCORD_WEBHOOK_URL', 'url')
+    @patch('freight.managers.FREIGHT_DISCORD_CUSTOMERS_WEBHOOK_URL', None)
+    @patch('freight.models.Webhook.execute', autospec=True)
+    def test_dont_send_pilot_notifications_for_expired_contracts(
+        self, 
+        mock_webhook_execute
+    ):        
+        x = Contract.objects.filter(status=Contract.STATUS_OUTSTANDING).first()
+        Contract.objects.all().exclude(pk=x.pk).delete()
+        x.date_expired = now() - datetime.timedelta(hours=1)
+        x.save()
+        self.assertTrue(tasks.send_contract_notifications(rate_limted=False))
+        self.assertEqual(mock_webhook_execute.call_count, 0)
+
+
+    @patch('freight.managers.FREIGHT_DISCORD_WEBHOOK_URL', None)
+    @patch('freight.managers.FREIGHT_DISCORD_CUSTOMERS_WEBHOOK_URL', 'url')
+    @patch('freight.models.Webhook.execute', autospec=True)
+    def test_dont_send_customer_notifications_for_expired_contracts(
+        self, 
+        mock_webhook_execute
+    ):        
+        x = Contract.objects.filter(status=Contract.STATUS_OUTSTANDING).first()
+        Contract.objects.all().exclude(pk=x.pk).delete()
+        x.date_expired = now() - datetime.timedelta(hours=1)
+        x.save()
+        self.assertTrue(tasks.send_contract_notifications(rate_limted=False))
+        self.assertEqual(mock_webhook_execute.call_count, 0)
+
+
+    @patch('freight.managers.FREIGHT_DISCORD_WEBHOOK_URL', 'url')
+    @patch('freight.managers.FREIGHT_DISCORD_CUSTOMERS_WEBHOOK_URL', None)
+    @patch('freight.models.Webhook.execute', autospec=True)
+    def test_send_pilot_notifications_only_once(
+        self, 
+        mock_webhook_execute
+    ):        
+        x = Contract.objects.filter(status=Contract.STATUS_OUTSTANDING).first()
+        Contract.objects.all().exclude(pk=x.pk).delete()        
+        
+        # round #1
+        self.assertTrue(tasks.send_contract_notifications(rate_limted=False))
+        self.assertEqual(mock_webhook_execute.call_count, 1)
+
+        # round #2
+        self.assertTrue(tasks.send_contract_notifications(rate_limted=False))
+        self.assertEqual(mock_webhook_execute.call_count, 1)
+
+
+    @patch('freight.managers.FREIGHT_DISCORD_WEBHOOK_URL', None)
+    @patch('freight.managers.FREIGHT_DISCORD_CUSTOMERS_WEBHOOK_URL', 'url')
+    @patch('freight.models.Webhook.execute', autospec=True)
+    def test_send_customer_notifications_only_once_per_state(
+        self, 
+        mock_webhook_execute
+    ):        
+        x = Contract.objects.filter(status=Contract.STATUS_OUTSTANDING).first()
+        Contract.objects.all().exclude(pk=x.pk).delete()        
+        
+        # round #1
+        self.assertTrue(tasks.send_contract_notifications(rate_limted=False))
+        self.assertEqual(mock_webhook_execute.call_count, 1)
+
+        # round #2
+        self.assertTrue(tasks.send_contract_notifications(rate_limted=False))
+        self.assertEqual(mock_webhook_execute.call_count, 1)
+
+
+    """
+    @patch('freight.managers.FREIGHT_DISCORD_WEBHOOK_URL', None)
+    @patch('freight.managers.FREIGHT_DISCORD_CUSTOMERS_WEBHOOK_URL', 'url')
+    @patch('freight.models.dhooks_lite.execute', autospec=True)
+    def test_dont_send_customer_notifications_for_expired_contracts(
+        self, 
+        mock_send_pilot_notification
+    ):                       
+        self.assertTrue(tasks.send_contract_notifications(rate_limted=False))
+        self.assertEqual(mock_send_pilot_notification.call_count, 0)
+
+    """